+++ .github/workflows/publish.yml
@@ -1,0 +1,68 @@
+name: publish
+
+on: 
+  workflow_dispatch
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: set up JDK 17
+        uses: actions/setup-java@v2
+        with:
+          distribution: 'adopt-hotspot'
+          java-version: 17
+          cache: 'gradle'
+
+      - name: make gradle wrapper executable
+        run: chmod +x ./gradlew
+      - name: build files
+        run: ./gradlew build
+
+      - name: read mod name, mod version and minecraft version
+        id: properties
+        uses: christian-draeger/read-properties@1.1.1
+        with:
+          path: './gradle.properties'
+          properties: 'modFilename mod_version modMinecraftVersion'
+
+      - uses: Kir-Antipov/mc-publish@v3.3
+        with:
+          modrinth-featured: true
+          modrinth-unfeature-mode: subset
+
+          modrinth-token: ${{ secrets.MODRINTH_TOKEN }}
+          curseforge-token: ${{ secrets.CURSEFORGE_TOKEN }}
+
+          name: '${{ steps.properties.outputs.modFilename }}-${{ steps.properties.outputs.mod_version }}+${{ steps.properties.outputs.modMinecraftVersion }}'
+          version: '${{ steps.properties.outputs.mod_version }}+${{ steps.properties.outputs.modMinecraftVersion }}'
+          game-versions: |
+            ${{ steps.properties.outputs.modMinecraftVersion }}
+
+          version-type: release
+          version-resolver: exact
+
+          dependencies: |
+            cloth-config | depends | *
+
+          changelog-file: changelog.md
+
+          retry-attempts: 2
+          retry-delay: 10000
+          fail-mode: fail
+
+      - name: read changelog
+        id: changelog
+        uses: juliangruber/read-file-action@v1
+        with:
+          path: 'changelog.md'
+
+      - name: discord webhook
+        uses: tsickert/discord-webhook@v5.3.0
+        with:
+          webhook-url: ${{ secrets.WEBHOOK_URL }}
+          embed-title: "${{ steps.properties.outputs.modFilename }} ${{ steps.properties.outputs.mod_version }} for Minecraft ${{ steps.properties.outputs.modMinecraftVersion }} has been released!"
+          embed-description: "**Changelog**```md\n${{ steps.changelog.outputs.content }}```\n**Downloads**\n<:modrinth:1110118061106798644> https://modrinth.com/mod/${{ steps.properties.outputs.modFilename }}\n<:curseforge:1110124796362109028> https://www.curseforge.com/minecraft/mc-mods/${{ steps.properties.outputs.modFilename }}"
+          embed-footer-text: "Have fun and be blessed!"
+          embed-color: 0xFF8000
