+++ gradle.properties
@@ -3,11 +3,11 @@
 org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
 
-modMinecraftVersion = 1.19.3
-modForgeVersion = 44.1.20
+modMinecraftVersion = 1.20.1
+modForgeVersion = 47.3.0
 modMixinVersion = 0.8.5
 mappingsChannel = parchment
-mappingsVersion = 2023.02.26-1.19.3
+mappingsVersion = 2023.08.20-1.20.1
 
 modId = rpgz
 modGroup = net.rpgz
@@ -17,7 +17,6 @@ modCurseId = 404828
 # Mod Properties
 	mod_version = 0.5.4-forge
 	maven_group = net.rpgz
-	archives_base_name = rpgz
 	
 # Dependencies
-clothConfigVersion = 9.0.94
+clothConfigVersion = 11.1.118
