+++ jitpack.yml
@@ -1,3 +1,3 @@
 before_install:
-  - wget https://github.com/sormuras/bach/raw/releases/11/install-jdk.sh
-  - source install-jdk.sh --feature 17
+  - sdk install java 17.0.3-tem
+  - sdk use java 17.0.3-tem
