+++ src/main/resources/META-INF/mods.toml
@@ -6,12 +6,12 @@ issueTrackerURL="https://github.com/Globox1997/RpgZ/issues"
 logoFile="icon.png"
 [[mods]]
 modId="rpgz"
-version="${file.jarVersion}"
+version="${mod_version}"
 displayName="RpgZ"
 updateJSONURL="https://raw.githubusercontent.com/Globox1997/RpgZ/master/update.json"
 displayURL="https://www.curseforge.com/minecraft/mc-mods/rpgz"
 credits="Obviously you for downloading this mod"
-authors="Globox_Z"
+authors="Globox_Z, Stereowalker"
 description='''
 Rpg Style Looting
 '''
@@ -22,7 +22,7 @@ Rpg Style Looting
     # Does this dependency have to exist - if not, ordering below must be specified
     mandatory=true #mandatory
     # The version range of the dependency
-    versionRange="[44,)" #mandatory
+    versionRange="[${modForgeVersion},)" #mandatory
     # An ordering relationship for the dependency - BEFORE or AFTER required if the relationship is not mandatory
     ordering="NONE"
     # Side this dependency is applied on - BOTH, CLIENT or SERVER
@@ -31,12 +31,12 @@ Rpg Style Looting
 [[dependencies.rpgz]]
     modId="minecraft"
     mandatory=true
-    versionRange="[1.19.3,1.20)"
+    versionRange="[${minecraft_version},)"
     ordering="NONE"
     side="BOTH"
 [[dependencies.rpgz]]
     modId="cloth_config"
     mandatory=true
-    versionRange="[9.0.94,)"
+    versionRange="[${clothConfigVersion},)"
     ordering="NONE"
     side="BOTH"
